+++ nbproject/private/private.properties
@@ -1,4 +1,7 @@
 deploy.ant.properties.file=C:\\Users\\FPT\\AppData\\Roaming\\NetBeans\\17\\tomcat101.properties
 j2ee.server.home=C:/Program Files/Apache Software Foundation/Tomcat 10.1
 j2ee.server.instance=tomcat101:home=C:\\Program Files\\Apache Software Foundation\\Tomcat 10.1
+javac.debug=true
+javadoc.preview=true
+selected.browser=default
 user.properties.file=C:\\Users\\FPT\\AppData\\Roaming\\NetBeans\\17\\build.properties
