+++ nbproject/private/private.xml
@@ -3,10 +3,8 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
-            <file>file:/D:/FptUniversity/25-SP/PRJ301/Assigment/Project/web/views/admin/listcourses.jsp</file>
-            <file>file:/D:/FptUniversity/25-SP/PRJ301/Assigment/Project/src/java/controller/admin/DeleteCourse.java</file>
-            <file>file:/D:/FptUniversity/25-SP/PRJ301/Assigment/Project/web/views/admin/addcourse.jsp</file>
-            <file>file:/D:/FptUniversity/25-SP/PRJ301/Assigment/Project/src/java/dal/CategoryDAO.java</file>
+            <file>file:/D:/FptUniversity/25-SP/PRJ301/Assigment/Group14_CourseSeller_Assignment/web/views/admin/listcourses.jsp</file>
+            <file>file:/D:/FptUniversity/25-SP/PRJ301/Assigment/Group14_CourseSeller_Assignment/web/views/admin/addcourse.jsp</file>
         </group>
     </open-files>
 </project-private>
