+++ nbproject/project.properties
@@ -25,11 +25,12 @@ dist.dir=dist
 dist.ear.war=${dist.dir}/${war.ear.name}
 dist.javadoc.dir=${dist.dir}/javadoc
 dist.war=${dist.dir}/${war.name}
+endorsed.classpath=
 excludes=
-file.reference.jakarta.servlet.jsp.jstl-2.0.0.jar=D:\\FptUniversity\\25-SP\\PRJ301\\lib\\jakarta.servlet.jsp.jstl-2.0.0.jar
-file.reference.jakarta.servlet.jsp.jstl-api-2.0.0.jar=D:\\FptUniversity\\25-SP\\PRJ301\\lib\\jakarta.servlet.jsp.jstl-api-2.0.0.jar
-file.reference.jaxb-api-2.1.jar=D:\\FptUniversity\\25-SP\\PRJ301\\lib\\jaxb-api-2.1.jar
-file.reference.sqljdbc42.jar=D:\\FptUniversity\\25-SP\\PRJ301\\lib\\sqljdbc42.jar
+file.reference.jakarta.servlet.jsp.jstl-2.0.0.jar=../../lib/jakarta.servlet.jsp.jstl-2.0.0.jar
+file.reference.jakarta.servlet.jsp.jstl-api-2.0.0.jar=../../lib/jakarta.servlet.jsp.jstl-api-2.0.0.jar
+file.reference.jaxb-api-2.1.jar=../../lib/jaxb-api-2.1.jar
+file.reference.sqljdbc42.jar=../../lib/sqljdbc42.jar
 includes=**
 j2ee.compile.on.save=true
 j2ee.copy.static.files.on.save=true
@@ -69,7 +70,6 @@ javadoc.use=true
 javadoc.version=false
 javadoc.windowtitle=
 lib.dir=${web.docbase.dir}/WEB-INF/lib
-no.dependencies=false
 persistence.xml.dir=${conf.dir}
 platform.active=default_platform
 resource.dir=setup
