+++ src/java/controller/admin/AddCourseForm.java
@@ -6,6 +6,7 @@
 package controller.admin;
 
 import dal.CategoryDAO;
+import dal.CourseDAO;
 import java.io.IOException;
 import java.io.PrintWriter;
 import jakarta.servlet.ServletException;
