+++ src/java/dal/CourseDAO.java
@@ -1,0 +1,96 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
+package dal;
+
+import java.sql.*;
+import java.sql.SQLException;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.logging.Level;
+import java.util.logging.Logger;
+import model.Category;
+import model.Courses;
+
+/**
+ *
+ * @author admin
+ */
+public class CourseDAO extends DBContext {
+
+    public List<Courses> getCourses() {
+        String sql = "SELECT *\n"
+                + "  FROM [dbo].[Courses]";
+        List<Courses> list = new ArrayList<>();
+        try {
+            PreparedStatement ps = connection.prepareStatement(sql);
+            ResultSet rs = ps.executeQuery();
+            while (rs.next()) {
+                Courses c = new Courses(rs.getInt("courseID"), rs.getString("title"), rs.getString("description"),
+                        rs.getInt("price"), rs.getInt("duration"), rs.getString("thumbnail"), rs.getString("createdDate"));
+                list.add(c);
+            }
+        } catch (SQLException ex) {
+            Logger.getLogger(CourseDAO.class.getName()).log(Level.SEVERE, null, ex);
+        }
+        return list;
+    }
+
+    private Connection getConnection() {
+        throw new UnsupportedOperationException("Not supported yet."); // Generated from nbfs://nbhost/SystemFileSystem/Templates/Classes/Code/GeneratedMethodBody
+    }
+
+    public void insertCourse(Courses c) {
+        String sql = "INSERT INTO Courses (Title, Description, Price, Duration, Thumbnail, CategoryID) VALUES (?, ?, ?, ?, ?, ?)";
+
+        try (PreparedStatement ps = connection.prepareStatement(sql)) {
+            ps.setString(1, c.getTitle());
+            ps.setString(2, c.getDescription());
+            ps.setInt(3, c.getPrice());
+            ps.setInt(4, c.getDuration());
+            ps.setString(5, c.getThumbnail());
+            ps.setInt(6, c.getCategoryID());
+            ps.executeUpdate();
+        } catch (SQLException ex) {
+            Logger.getLogger(CourseDAO.class.getName()).log(Level.SEVERE, null, ex);
+        }
+    }
+
+    public void deleteCourse(int id) {
+        String sql = "DELETE FROM CourseVideos WHERE CourseID = ?;\n"
+                + "DELETE FROM OrderDetails WHERE CourseID = ?;\n"
+                + "DELETE FROM Reviews WHERE CourseID = ?;\n"
+                + "DELETE FROM Courses WHERE CourseID = ?;";
+
+        try (PreparedStatement ps = connection.prepareStatement(sql)) {
+            ps.setInt(1, id);
+            ps.setInt(2, id);
+            ps.setInt(3, id);
+            ps.setInt(4, id);
+            ps.executeUpdate();
+        } catch (SQLException ex) {
+            Logger.getLogger(CourseDAO.class.getName()).log(Level.SEVERE, null, ex);
+        }
+    }
+
+    public Courses getCourseByTitle(String title) {
+        String sql = "SELECT *\n"
+                + "  FROM [dbo].[Courses]"
+                + " WHERE Title = ?";
+        try {
+            PreparedStatement ps = connection.prepareStatement(sql);
+            ps.setString(1, title);
+            ResultSet rs = ps.executeQuery();
+            if (rs.next()) {
+                Courses c = new Courses(rs.getInt("courseID"), rs.getString("title"), rs.getString("description"),
+                        rs.getInt("price"), rs.getInt("duration"), rs.getString("thumbnail"), rs.getInt("categoryID"), rs.getString("createdDate"));
+                return c;
+            }
+        } catch (SQLException ex) {
+            System.out.println(ex.getMessage());
+        }
+        return null;
+    }
+
+}
